pipeline fix for uneven number of scene channel files
